+++ superwall/src/test/java/com/superwall/sdk/paywall/view/PaywallViewTest.kt
@@ -21,7 +21,6 @@ import com.superwall.sdk.paywall.view.webview.PaywallMessage
 import com.superwall.sdk.paywall.view.webview.PaywallUIDelegate
 import com.superwall.sdk.paywall.view.webview.PaywallWebUI
 import com.superwall.sdk.paywall.view.webview.SendPaywallMessages
-import com.superwall.sdk.paywall.view.webview.WebviewError
 import com.superwall.sdk.paywall.view.webview.messaging.PaywallMessageHandler
 import com.superwall.sdk.paywall.view.webview.messaging.PaywallWebEvent
 import com.superwall.sdk.paywall.view.webview.templating.models.JsonVariables
@@ -511,7 +510,6 @@ class PaywallViewTest {
         override fun setup(
             url: com.superwall.sdk.models.paywall.PaywallURL,
             onRenderCrashed: (Boolean, Int) -> Unit,
-            onTimeout: ((WebviewError) -> Unit)?,
         ) {
             lastSetupUrl = url
             setupLatch?.countDown()